:recycle: Move section controls to own component
Split rendering of p node from section node
